+++ static/js/nzbhydra.js
@@ -3447,7 +3447,7 @@ function SearchController($scope, $http, $stateParams, $state, $window, $filter,
 
     function getAndSetSearchRequests() {
         SearchHistoryService.getSearchHistoryForSearching().success(function (data) {
-            $scope.searchHistory = data.data.searchRequests;
+            $scope.searchHistory = data.searchRequests;
         });
     }
 
@@ -5083,7 +5083,7 @@ function ConfigFields($injector) {
                                 help: 'Set when using an external proxy. Call using a trailing slash, e.g. http://www.domain.com/nzbhydra/'
                             },
                             validators: {
-                                urlBase: regexValidator(/^\/[\w\/]*$/, "Base URL needs to start with a slash and must not end with one")
+                                urlBase: regexValidator(/^(\/\w+)*$/, "Base URL needs to start with a slash and must not end with one")
                             }
                         },
                         {